Chore: 共通処理部分を統合
