+++ Assets/Project/Scripts/Infra/StageDataObject.cs
@@ -17,5 +17,10 @@ namespace Project.Scripts.Infra
         public int stageNumber;  // 文字だと扱いづらいので、int型のステージ番号も持つ
         public string charaStillAddress;
         public string title;
+        [Header("Battle Flow")]
+        public string waySequenceAddress;
+        public string bossSequenceAddress;
+        public string scenarioIdWayToBoss;
+        public string scenarioIdBossToNext;
     }
 }
