--- Assets/Project/Scripts/Model/RuntimeModel-battle.cs
@@ -1,7 +1,0 @@
-﻿namespace Project.Scripts.Model
-{
-    public class RuntimeModel : ModelBase
-    {
-        public int CurrentStageNumber { get; internal set; } = -1;
-    }
-}
