--- Assets/Project/Scripts/Model/RuntimeModel-battle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 63912aa620987aa43be7951eba5a4477
