--- Assets/Project/Scripts/Model/RuntimeModel-scenario.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cfbfe8e87fe821d4fb0fc86b4b769d6f
