--- Assets/Project/Scripts/Model/StageModel-battle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 643494f66b31e9e4fa2d9e20e8c0a069
