--- Assets/Project/Scripts/Model/StageModel-scenario.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0316fad7af909de4ea15b5d142122618
